+++ lib/hyrax/doi.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 require "hyrax/doi/engine"
 require "hyrax/doi/errors"
+require "hyrax/doi/configuration"
 require "flipflop"
 
 module Hyrax
   module DOI
-    # Your code goes here...
   end
 end